+++ admin/src/main/kotlin/com/nexters/admin/AdminApplication.kt
@@ -6,7 +6,7 @@ import org.springframework.context.annotation.ComponentScan
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories
 
 @SpringBootApplication
-@ComponentScan(basePackages = ["com.nexters.admin", "com.nexters.external"])
+@ComponentScan(basePackages = ["com.nexters.admin", "com.nexters.external", "com.nexters.newsletter"])
 @EnableJpaRepositories(basePackages = ["com.nexters.admin.repository"])
 class AdminApplication
