--- admin/src/main/kotlin/com/nexters/admin/config/AdminMemberInitializer.kt
@@ -1,1 +1,0 @@
-// 이 파일은 삭제됨 - 관리자는 수동으로 DB에 직접 추가
