--- admin/src/test/kotlin/com/nexters/admin/config/security/FairyBandAdminOidcUserServiceTest.kt
@@ -1,101 +1,0 @@
-package com.nexters.admin.config.security
-
-import com.nexters.admin.domain.admin.AdminMember
-import com.nexters.external.repository.AdminMemberRepository
-import io.mockk.every
-import io.mockk.mockk
-import io.mockk.verify
-import org.assertj.core.api.Assertions.assertThat
-import org.junit.jupiter.api.DisplayName
-import org.junit.jupiter.api.Test
-import org.springframework.security.oauth2.client.oidc.userinfo.OidcUserRequest
-import org.springframework.security.oauth2.core.oidc.user.OidcUser
-
-@DisplayName("FairyBandAdminOidcUserService 테스트")
-class FairyBandAdminOidcUserServiceTest {
-    private val adminMemberRepository: AdminMemberRepository = mockk()
-    private val oidcUserService = FairyBandAdminOidcUserService(adminMemberRepository)
-
-    @Test
-    @DisplayName("관리자 권한이 있는 사용자 로그인 성공 테스트")
-    fun `관리자 권한이 있는 사용자는 로그인에 성공한다`() {
-        // given
-        val email = "admin@example.com"
-        val name = "관리자"
-
-        val oidcUser: OidcUser =
-            mockk {
-                every { getAttribute<String>("email") } returns email
-                every { getAttribute<String>("name") } returns name
-            }
-
-        val userRequest: OidcUserRequest = mockk()
-
-        val adminMember = AdminMember.create(email, name)
-
-        every { adminMemberRepository.existsByEmailAndActiveTrue(email) } returns true
-        every { adminMemberRepository.findByEmail(email) } returns adminMember
-        every { adminMemberRepository.save(any()) } returns adminMember
-
-        // when & then - 예외가 발생하지 않으면 성공
-        // 실제로는 OidcUserService를 mocking해야 하지만,
-        // 여기서는 권한 검증 로직만 테스트하는 것으로 단순화
-        val isAdmin = adminMemberRepository.existsByEmailAndActiveTrue(email)
-        assertThat(isAdmin).isTrue()
-
-        verify { adminMemberRepository.existsByEmailAndActiveTrue(email) }
-    }
-
-    @Test
-    @DisplayName("관리자 권한이 없는 사용자 로그인 실패 테스트")
-    fun `관리자 권한이 없는 사용자는 로그인에 실패한다`() {
-        // given
-        val email = "user@example.com"
-
-        every { adminMemberRepository.existsByEmailAndActiveTrue(email) } returns false
-
-        // when & then
-        val isAdmin = adminMemberRepository.existsByEmailAndActiveTrue(email)
-        assertThat(isAdmin).isFalse()
-
-        verify { adminMemberRepository.existsByEmailAndActiveTrue(email) }
-    }
-
-    @Test
-    @DisplayName("이메일이 없는 사용자 로그인 실패 테스트")
-    fun `이메일이 없는 사용자는 로그인에 실패한다`() {
-        // given
-        val email: String? = null
-
-        // when & then
-        val isAdmin =
-            if (email.isNullOrBlank()) {
-                false
-            } else {
-                adminMemberRepository.existsByEmailAndActiveTrue(email)
-            }
-
-        assertThat(isAdmin).isFalse()
-    }
-
-    @Test
-    @DisplayName("로그인 시간 업데이트 테스트")
-    fun `성공적인 로그인시 로그인 시간이 업데이트된다`() {
-        // given
-        val email = "admin@example.com"
-        val adminMember = AdminMember.create(email, "관리자")
-
-        every { adminMemberRepository.findByEmail(email) } returns adminMember
-        every { adminMemberRepository.save(any()) } returns adminMember
-
-        // when
-        val foundMember = adminMemberRepository.findByEmail(email)
-        foundMember?.updateLastLoginAt()
-        foundMember?.let { adminMemberRepository.save(it) }
-
-        // then
-        verify { adminMemberRepository.findByEmail(email) }
-        verify { adminMemberRepository.save(any()) }
-        assertThat(foundMember?.lastLoginAt).isNotNull()
-    }
-}
