+++ api/build.gradle.kts
@@ -1,5 +1,5 @@
 dependencies {
-    implementation(project(":external"))
+    implementation(project(":newsletter"))
     implementation(libs.spring.boot.starter.web)
     implementation(libs.spring.boot.starter.actuator)
     implementation(libs.jackson.module.kotlin)
