+++ api/src/main/kotlin/com/nexters/api/controller/NewsletterApiController.kt
@@ -2,7 +2,7 @@ package com.nexters.api.controller
 
 import com.nexters.api.dto.ContentViewApiResponse
 import com.nexters.api.dto.ContentViewApiResponse.ContentCardApiResponse
-import com.nexters.external.resolver.DailyContentArchiveResolver
+import com.nexters.newsletter.resolver.DailyContentArchiveResolver
 import io.swagger.v3.oas.annotations.tags.Tag
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
@@ -14,12 +14,12 @@ import java.time.LocalDate
 @RequestMapping("/api/newsletters")
 @Tag(name = "Newsletter API", description = "뉴스레터 관련 API")
 class NewsletterApiController(
-    private val dayArchiveResolver: DailyContentArchiveResolver
+    private val dayArchiveResolver: DailyContentArchiveResolver,
 ) {
     @GetMapping("/contents/{userId}")
     fun getNewsletterContents(
         @PathVariable userId: Long,
-        publishedDate: LocalDate = LocalDate.now()
+        publishedDate: LocalDate = LocalDate.now(),
     ): ContentViewApiResponse =
         ContentViewApiResponse(
             publishedDate = publishedDate,
@@ -30,8 +30,8 @@ class NewsletterApiController(
                         topKeyword = it.provocativeKeyword,
                         summary = it.summaryContent,
                         contentUrl = it.content.originalUrl,
-                        newsletterName = it.content.newsletterName
+                        newsletterName = it.content.newsletterName,
                     )
-                }
+                },
         )
 }
