--- external/src/main/kotlin/com/nexters/external/parser/MailContent.kt
@@ -1,8 +1,0 @@
-package com.nexters.external.parser
-
-data class MailContent(
-    val title: String,
-    val content: String,
-    val link: String,
-    val section: String? = null,
-)
