--- external/src/main/kotlin/com/nexters/external/parser/MailParser.kt
@@ -1,7 +1,0 @@
-package com.nexters.external.parser
-
-interface MailParser {
-    fun isTarget(sender: String): Boolean
-
-    fun parse(content: String): List<MailContent>
-}
