--- external/src/main/kotlin/com/nexters/external/parser/MailParserFactory.kt
@@ -1,16 +1,0 @@
-package com.nexters.external.parser
-
-class MailParserFactory {
-    private val parsers =
-        listOf(
-            KotlinWeeklyParser(),
-            CssWeeklyParser(),
-            SwiftVincentParser(),
-            ReactStatusParser(),
-            IlbunParser()
-        )
-
-    fun findParser(sender: String): MailParser? = parsers.find { it.isTarget(sender) }
-
-    fun getAllParsers(): List<MailParser> = parsers
-}
