+++ newsletter/src/main/kotlin/com/nexters/newsletter/parser/MailContent.kt
@@ -1,0 +1,8 @@
+package com.nexters.newsletter.parser
+
+data class MailContent(
+    val title: String,
+    val content: String,
+    val link: String,
+    val section: String? = null,
+)
