+++ newsletter/src/main/kotlin/com/nexters/newsletter/parser/MailParser.kt
@@ -1,0 +1,7 @@
+package com.nexters.newsletter.parser
+
+sealed interface MailParser {
+    fun isTarget(sender: String): Boolean
+
+    fun parse(content: String): List<MailContent>
+}
