+++ newsletter/src/main/kotlin/com/nexters/newsletter/parser/MailParserFactory.kt
@@ -1,0 +1,16 @@
+package com.nexters.newsletter.parser
+
+class MailParserFactory {
+    private val parsers =
+        listOf(
+            KotlinWeeklyParser(),
+            CssWeeklyParser(),
+            SwiftVincentParser(),
+            ReactStatusParser(),
+            IlbunParser()
+        )
+
+    fun findParser(sender: String): MailParser? = parsers.find { it.isTarget(sender) }
+
+    fun getAllParsers(): List<MailParser> = parsers
+}
