+++ newsletter/src/test/resources/ilbun-html-part.html
