+++ newsletter/src/test/resources/swift-vincent
