+++ settings.gradle.kts
@@ -16,3 +16,4 @@ include("api")
 include("batch")
 include("external")
 include("admin")
+include("newsletter")
